// type.d.ts
export type ProductProps = {
  brand: string;
  category: string;
  description: string;
  image: string;
  isNew: boolean;
  oldPrice: number;
  price: number;
  title: string;
  _id: number;
};

export type StoreProduct = {
  brand: string;
  category: string;
  description: string;
  image: string;
  isNew: boolean;
  oldPrice: number;
  price: number;
  title: string;
  _id: number;
  quantity: number;
};

export type StateProps = {
  productData: [];
  favoriteData: [];
  userInfo: null | string;
  next: any;
};